backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
pydantic>=2.5.0
python-multipart>=0.0.6
openpyxl>=3.1.2
pandas>=2.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
aiofiles>=23.2.1
jinja2>=3.1.2
reportlab>=4.0.7
PyJWT>=2.8.0
requests>=2.31.0
openai>=1.0.0 